third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/style-of-distributed-node-expected.txt

CONSOLE ERROR: line 27: Uncaught TypeError: document.getElementById(...).createShadowRoot is not a function
Tests to ensure that styles of distributed nodes are different if their parent styles are different.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




    

        

        

    







FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
